# Communicator/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Communicator",
    "description": "Email Notifications Sender",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Full Subscription",
      "description": "Event is fully booked"
    },
    {
      "name": "Event Updates",
      "description": "Event is updated/cancelled"
    },
    {
      "name": "Upcoming Event Subscription",
      "description": "Event is upcoming"
    }
  ],
  "paths": {
    "/events/fullSubscription": {
      "post": {
        "tags": [
          "Full Subscription"
        ],
        "summary": "Subscribe user to event notification",
        "description": "Lorem Ipsum Dolor Sit Amet",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "type": "integer",
            "description": "The event ID"
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Full Subscription"
        ],
        "summary": "Unsubscribe user from event notification",
        "description": "Lorem Ipsum Dolor Sit Amet",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "type": "integer",
            "description": "The event ID"
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events/eventUpdates": {
      "post": {
        "tags": [
          "Event Updates"
        ],
        "summary": "Subscribe user to event notification",
        "description": "This allows the user to receive an email whenever a change is made by the admin to an event. When the event is cancelled, this sends out a cancellation email to all volunteers.",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "type": "integer",
            "description": "The event ID"
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Event Updates"
        ],
        "summary": "Unsubscribe user from event notification",
        "description": "This allows the user to receive an email whenever a change is made by the admin to an event. When the event is cancelled, this sends out a cancellation email to all volunteers.",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "type": "integer",
            "description": "The event ID"
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events/upcomingEventSubscription": {
      "post": {
        "tags": [
          "Upcoming Event Subscription"
        ],
        "summary": "Subscribe user to event notification",
        "description": "This sends a reminder email to all participants stating that the event will happen the following day.",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "type": "integer",
            "description": "The event ID"
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Upcoming Event Subscription"
        ],
        "summary": "Unsubscribe user from event notification",
        "description": "This sends a reminder email to all participants stating that the event will happen the following day.",
        "parameters": [
          {
            "in": "query",
            "name": "eventId",
            "required": true,
            "type": "integer",
            "description": "The event ID"
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}